Change find_xy_shell to use astropy.coordinates
Ditch the home-made parsing of the RA, Dec in favor of using the
functionality of astropy.coordinates.   This affects the initial
calculation of the x and y of the shells.   The remaining part (finding
shell axis, etc) is unchanged.   We use a new modules angle_utils.py,
which has a function for working out the vector separation between two
sky coordinates.