Tail fits are huch better now
I tried to do a different matching strategy.  This is not working yet,
but in the meantime I have found that the better approx for th1,90 means
that the original strategy now works much better.  We still have to
chose the position to do the head/tail matching, x_m, but the fits are
now very good for all models except the k=0 anisotropic, which has a
discontinuity in the second derivative at x=0