Investigate residuals for multi-level analytic tail fits
Everything looks OK.  There is a slight issue with big
residuals (around the 10% level) with the very low beta models for
theta > 170, which is because that range of theta was not included in
the fit optimizations.  But we can't do anything about that without
making the fit far worse in the more important range of 110 -> 160
degrees.  So, we will just have to live with it.

Next thing to do is to make a multi-panel plot showing these residuals
for all (or maybe just some?) of the parameter sets used in Figure
10.