Update fit-circle-shell.py for python 3 and latest lmfit
The most important fix is to deal with a change in the API of lmfit.
Recent versions do not overwrite the initial parameters, but return a
structure with the optimization result

I have also added a new command line argument, --thmax, for specifying
the maximum angle from the symmetry axis to use in the circle fits.

Finally, there are some minor changes for compatibility with python
3.  Note that changes to fit-conic-shell.py and conic_utils.py are
still incomplete.