Yet more fiddling with big p-value table
I decided to reverse the sign of Cohen's d for consistency with the
rank biserial correlation coefficient.  Also turn the sig/tau value
upside down so that it is usually bigger than one.  The table is now
generated automatically from the JSON file, so any further adjustments
will be easy.