Improve the smoothness of the dragoid shapes near the apex
There was an instability in the determination of R_c for the dragoids,
in that it was very dependent on the delta theta used and the order of
polynomial for the fit.  It turns out that this is mainly due to noise
in the shape for low values of theta (< 15 deg).  So I now do a lowess
smoothing of the dragoid R(theta), but apply it wit a weighting
functions that falls as a gaussian of theta with sigma=45 deg.  So it
has almost no effect beyond 60 deg (which we need).  This works pretty
well and means that we can now use the same values of
SCALE_NEIGHBORHOOD and DEGREE_POLY_NEIGHBORHOOD that we were using for
the ancantoids.