Finish initial test of the simulation shape extrapolation
We can either take the positive or the negative values of theta, and
it makes a slight difference for the MHD case, suggesting that we need
to make sure that the fit has PA=0, which it doesn't at the moment.

I also tried folding together the positive and negative values, but
even using a bit of lowess smoothing, we get some visible wiggles in
the spline fit, especially around theta = 150 deg.

One solution might be to use a Chebyshev fit to Delta(mu) instead of
the splines.